+++ CHANGELOG.md
@@ -10,3 +10,13 @@ This document contains a list of all the changes for each version of Marksman.
 March 22nd, 2023
 
 - Core functionality
+
+
+## Version 1.0
+
+### Name Pending
+
+*Release date to be determined*
+
+- Added refresh delay configuration option.
+- Add question mark icon to replace directional icons when there isn't enough information to display accurate data.